requirements.txt
# LangChain Core
langchain
langchain-core
langchain_community
langchain_classic
langchain-huggingface

# Embeddings
sentence-transformers

# Groq
langchain-groq

# Google Gemini
langchain-google-genai
google-generativeai

# Vector DB
chromadb

# Environment
python-dotenv

# UI
streamlit

# PDF & Document Processing
pypdf
pymupdf
unstructured
pdf2image
pdfminer.six
unstructured_pytesseract
unstructured_inference
